Replace nonnegative amount with u64, start on ValueTransferSummary type and helpers

+++ zingolib/src/blaze/fetch_full_transaction.rs
@@ -154,7 +154,7 @@ impl TransactionContext {
                     {
                         outgoing_metadatas.push(OutgoingTxData {
                             to_address: taddr,
-                            value: vout.value,
+                            value: u64::from(vout.value),
                             memo: Memo::Empty,
                             recipient_ua: None,
                         });
